test(expect): move type-only tests to jest-expect.test-d.ts
